test(FSADT1-1452): moving staff registered to e2e
