+++ frontend/cypress/fixtures/clients/bcreg_ac_list4.json
@@ -1,0 +1,26 @@
+[
+  {
+    "code": "C1234567",
+    "name": "Existing Corporation 1",
+    "status": "ACTIVE",
+    "legalType": "C"
+  },
+  {
+    "code": "C1231231",
+    "name": "Existing Corporation with a super long name that should not be allowed here",
+    "status": "ACTIVE",
+    "legalType": "C"
+  },
+  {
+    "code": "FM123456",
+    "name": "Existing Sole Proprietorship 1",
+    "status": "ACTIVE",
+    "legalType": "SP"
+  },
+  {
+    "code": "FM999457",
+    "name": "Koalllá",
+    "status": "ACTIVE",
+    "legalType": "SP"
+  }
+]
