+++ frontend/cypress/fixtures/staff/bcregisteredscenarios.json
@@ -1,0 +1,133 @@
+[
+  {
+    "scenarioName": "Failed state - Duplicated entry",
+    "companySearch": "dup",
+    "companyCode": "C7775745",
+    "showData": true,
+    "showBirthdate": false,
+    "showUnknowNotification": true,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": false,
+    "showDuplicatedNotification": true,
+    "showNotOwnedByPersonError": false,
+    "type": "Corporation",
+    "standing": "Unknown",
+    "dba": ""
+  },
+  {
+    "scenarioName": "OK State - Sole Proprietorship, show birthdate",
+    "companySearch": "spp",
+    "companyCode": "FM123123",
+    "showData": true,
+    "showBirthdate": true,
+    "showUnknowNotification": false,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": false,
+    "showDuplicatedNotification": false,
+    "showNotOwnedByPersonError": false,
+    "type": "Sole proprietorship",
+    "standing": "Good standing",
+    "dba": "Soleprop"
+  },
+  {
+    "scenarioName": "Failed state - BC Registry down",
+    "companySearch": "bcd",
+    "companyCode": "C7745745",
+    "showData": false,
+    "showBirthdate": false,
+    "showUnknowNotification": false,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": true,
+    "showDuplicatedNotification": false,
+    "showNotOwnedByPersonError": false,
+    "type": "Corporation",
+    "standing": "Good Standing",
+    "dba": ""
+  },      
+  {
+    "scenarioName": "Failed state - Not in good standing",
+    "companySearch": "ngs",
+    "companyCode": "C4443332",
+    "showData": true,
+    "showBirthdate": false,
+    "showUnknowNotification": false,
+    "showNotGoodStandingNotification": true,
+    "showBcRegDownNotification": false,
+    "type": "Corporation",
+    "standing": "Not in good standing",
+    "dba": ""
+  },
+  {
+    "scenarioName": "OK State - Corporation",
+    "companySearch": "cmp",
+    "companyCode": "C1231231",
+    "showData": true,
+    "showBirthdate": false,
+    "showUnknowNotification": false,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": false,
+    "showDuplicatedNotification": false,
+    "showNotOwnedByPersonError": false,
+    "type": "Corporation",
+    "standing": "Good standing",
+    "dba": ""
+  },
+  {
+    "scenarioName": "OK State - Corporation not found",
+    "companySearch": "cnf",
+    "companyCode": "C9999999",
+    "showData": true,
+    "showBirthdate": false,
+    "showUnknowNotification": true,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": false,
+    "showDuplicatedNotification": false,
+    "showNotOwnedByPersonError": false,
+    "type": "Corporation",
+    "standing": "Unknown",
+    "dba": ""
+  },      
+  {
+    "scenarioName": "OK State - Unsuported types for external",
+    "companySearch": "llp",
+    "companyCode": "LL123123",
+    "showData": false,
+    "showBirthdate": false,
+    "showUnknowNotification": false,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": false,
+    "type": "Limited liability partnership",
+    "standing": "Unknown",
+    "dba": ""
+  },
+  {
+    "scenarioName": "Failed state - Unknown standing",
+    "companySearch": "uks",
+    "companyCode": "C4566541",
+    "showData": true,
+    "showBirthdate": false,
+    "showUnknowNotification": true,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": false,
+    "showDuplicatedNotification": false,
+    "showNotOwnedByPersonError": false,
+    "type": "Corporation",
+    "standing": "Unknown",
+    "dba": ""
+  },
+  {
+    "scenarioName": "Failed state - SP not owned by person",
+    "companySearch": "spw",
+    "companyCode": "FM7715744",
+    "showData": true,
+    "showBirthdate": true,
+    "showUnknowNotification": false,
+    "showNotGoodStandingNotification": false,
+    "showBcRegDownNotification": false,
+    "showDuplicatedNotification": false,
+    "showNotOwnedByPersonError": true,
+    "type": "Sole proprietorship",
+    "standing": "Good standing",
+    "dba": "Soleprop"
+  }
+]
